create relationship between user and task

+++ task-app/src/routers/task.js
@@ -1,9 +1,13 @@
 const express = require('express');
 const Task = require('../models/task');
 const router = new express.Router();
+const auth = require('../middleware/auth');
 
-router.post('/tasks', (req, res) => {
-    const task = new Task(req.body)
+router.post('/tasks', auth,(req, res) => {
+    const task = new Task({
+        ...req.body,
+        owner: req.user._id
+    })
 
     task.save().then(() => {
         res.send(task)
@@ -13,29 +17,31 @@ router.post('/tasks', (req, res) => {
     })
 })
 
-router.get('/tasks', (req, res) => {
-    Task.find().then((tasks) => {
-        res.status(200).send(tasks)
-    }).catch((err) => {
+router.get('/tasks',auth, async(req, res) => {
+    try{
+        await req.user.populate('tasks').execPolpulate()
+        res.status(200).send(req.user.tasks)
+    }catch(err) {
         res.status(500).send(err)
-    })
+    }
 })
 
 //fetch particular field using url
-router.get('/tasks/:id', (req, res) => {
+router.get('/tasks/:id',auth, async(req, res) => {
     const _id = req.params.id
-    Task.findById(_id).then((task) => {
+    try{
+        const task = await Task.findOne({_id,owner: req.user._id})
         if(!task){
             return res.status(404).send()
         }
         res.send(task)
-    }).catch((e) => {
+    }catch(e) {
         res.status(500).send()
-    })
+    }
 })
 
 //we use async await
-router.patch("/tasks/:id", async (req, res) => {
+router.patch("/tasks/:id", auth ,async (req, res) => {
     const updates = Object.keys(req.body);
     const allowUpdates = ["description", "completed"];
     const isValidOperation = updates.every((update) =>
@@ -47,22 +53,26 @@ router.patch("/tasks/:id", async (req, res) => {
     }
 
     try {
-        const task = await Task.findByIdAndUpdate(req.params.id, req.body, {
-            new: true,
-            runValidators: true,
-        });
+        // const task = await Task.findByIdAndUpdate(req.params.id, req.body, {
+        //     new: true,
+        //     runValidators: true,
+        // });
+        const task = await Task.findOneAndUpdate({_id, owner: req.user._id})
         if (!task) {
             return res.status(404).send();
         }
+
+        updates.forEach((update)=>  task[update] = req.body[update])
+        await task.save(); 
         res.send(task);
     } catch (err) {
         res.status(500).send(e);
     }
 });
 
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id', auth ,async (req, res) => {
     try {
-        const task = await Task.findByIdAndDelete(req.params.id)
+        const task = await Task.findByIdAndDelete({_id:req.params.id , owner: req.user._id})
         if (!task) {
             return res.status(404).send();
         }
